+++ R/plot.R
@@ -50,9 +50,9 @@
 }
 
 # Private
-.drawTrajExtras <- function(x, draw.start.pt = TRUE, start.pt.cex = .8, turning.angles = NULL, ...) {
+.drawTrajExtras <- function(x, draw.start.pt = TRUE, start.pt.cex = .8, start.pt.pch = 16, start.pt.col = "black", turning.angles = NULL, ...) {
   if (draw.start.pt) {
-    graphics::points(x$x[1], x$y[1], pch = 16, cex = start.pt.cex)
+    graphics::points(x$x[1], x$y[1], pch = start.pt.pch, cex = start.pt.cex, col = start.pt.col)
   }
 
   if (!is.null(turning.angles)) {
@@ -72,6 +72,9 @@
 #' @param draw.start.pt If TRUE, draws a dot at the start point of the
 #'   trajectory.
 #' @param start.pt.cex Scale to apply when drawing the start point dot.
+#' @param start.pt.pch Pch (i.e. plot character, symbol or shape) to apply when
+#'   drawing the start point dot.
+#' @param start.pt.col Colour to apply when drawing the start point dot.
 #' @param turning.angles If \code{random} or \code{directed}, draws step turning
 #'   angles. \code{directed} assumes errors are relative to the first recorded
 #'   step angle. \code{random} assumes errors are relative to the previous step.
@@ -88,7 +91,8 @@
 #'
 #' @export
 plot.Trajectory <- function(x, add = FALSE,
-                            draw.start.pt = TRUE, start.pt.cex = TRUE, turning.angles = NULL,
+                            draw.start.pt = TRUE, start.pt.cex = 0.8, start.pt.pch = 16, start.pt.col = "black",
+                            turning.angles = NULL,
                             xlim = grDevices::extendrange(x$x), ylim = grDevices::extendrange(x$y),
                             xlab = ifelse(is.null(TrajGetUnits(x)), "x", sprintf("x (%s)", TrajGetUnits(x))),
                             ylab = ifelse(is.null(TrajGetUnits(x)), "y", sprintf("y (%s)", TrajGetUnits(x))),
@@ -96,7 +100,9 @@ plot.Trajectory <- function(x, add = FALSE,
   if (!add) {
     graphics::plot(NULL, xlim = xlim, ylim = ylim, xlab = xlab, ylab = ylab, asp = asp, ...)
   }
-  graphics::lines(x, draw.start.pt = draw.start.pt, start.pt.cex = start.pt.cex, turning.angles = turning.angles, ...)
+  graphics::lines(x, draw.start.pt = draw.start.pt,
+                  start.pt.cex = start.pt.cex, start.pt.pch = start.pt.pch, start.pt.col = start.pt.col,
+                  turning.angles = turning.angles, ...)
 }
 
 #' Add Trajectory lines to a plot
@@ -107,15 +113,18 @@ plot.Trajectory <- function(x, add = FALSE,
 #' @param draw.start.pt If TRUE, draws a dot at the start point of the
 #'   trajectory.
 #' @param start.pt.cex Scale to apply when drawing the start point dot.
+#' @param start.pt.pch Pch (i.e. plot character, symbol or shape) to apply when
+#'   drawing the start point dot.
+#' @param start.pt.col Colour to apply when drawing the start point dot.
 #' @param turning.angles If \code{random} or \code{directed}, draws step turning
 #'   angles. \code{directed} assumes errors are relative to the first recorded
 #'   step angle. \code{random} assumes errors are relative to the previous step.
 #' @param ... Additional arguments are passed to \code{\link[graphics]{lines}}.
 #'
 #' @export
-lines.Trajectory <- function(x, draw.start.pt = TRUE, start.pt.cex = 0.8, turning.angles = NULL, ...) {
+lines.Trajectory <- function(x, draw.start.pt = TRUE, start.pt.cex = 0.8, start.pt.pch = 16, start.pt.col = "black", turning.angles = NULL, ...) {
   graphics::lines(y ~ x, data = x, ...)
-  .drawTrajExtras(x, draw.start.pt, start.pt.cex, turning.angles, ...)
+  .drawTrajExtras(x, draw.start.pt, start.pt.cex, start.pt.pch = start.pt.pch, start.pt.col = start.pt.col, turning.angles, ...)
 }
 
 #' Add Trajectory points to a plot